backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
PyMySQL==1.1.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
